feat: reorganize donations with tabs
