+++ src/components/OtherDonations.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const OtherDonations = () => (
+  <div className="text-center py-6 space-y-2">
+    <h2 className="text-3xl font-bold text-foreground mb-2">Autres dons</h2>
+    <p className="text-muted-foreground">
+      Ajoutez ici les dons spécifiques bénéficiant d'une déduction plus avantageuse.
+    </p>
+  </div>
+);
+
+export default OtherDonations;
